Full compile
