+++ src/endpoints/grpc/types.h
@@ -13,22 +13,6 @@
 
 namespace ccf::grpc
 {
-  static const http::HeaderMap default_response_headers = {
-    {http::headers::CONTENT_TYPE, http::headervalues::contenttype::GRPC}};
-
-  static constexpr auto TRAILER_STATUS = "grpc-status";
-  static constexpr auto TRAILER_MESSAGE = "grpc-message";
-
-  static http::HeaderKeyValue make_status_trailer(int32_t code)
-  {
-    return {TRAILER_STATUS, std::to_string(code)};
-  }
-
-  static http::HeaderKeyValue make_message_trailer(const std::string& msg)
-  {
-    return {TRAILER_MESSAGE, msg};
-  }
-
   template <typename T>
   struct SuccessResponse
   {
